+++ src/Server/makeServer.ts
@@ -61,6 +61,30 @@ const makeServer = async (deps: Dependencies) => {
         })
     })
 
+    app.post('/api/workspaces/delete-by-path', async (req, res) => {
+        // Attempt to get the ID of the workspace
+        const existingWorkspaces = await deps.collection('workspaces').getAll()
+
+        // If not found, return a 404
+        const workspace = existingWorkspaces.filter(workspace => workspace.path === req.body.path)
+
+        if (workspace.length === 0) {
+            res.status(404)
+            res.json({
+                message: 'Workspace not found.'
+            })
+            return
+        }
+
+        // Otherwise, delete it
+        await deps.collection('workspaces').delete(workspace[0].id)
+
+        res.json({
+            message: 'Workspace successfully deleted.'
+        })
+
+    })
+
     app.all('/api/*', (req, res) => {
         res.sendStatus(404)
     })