+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - "at" style for links in documentation navigation
 - initial implementation of `jpx init` command
 - `jpx register` command
+- `jpx deregister` command
 - Project internals
     - Added debug timer to track server startup time
     - WebSocket route helpers